requirements.txt
asyncio>=3.4.3
asyncpg>=0.30.0
kafka-python>=2.1.5
pip-chill>=1.0.3
plotly>=6.0.1
psycopg2-binary>=2.9.10
pytest-asyncio>=0.26.0
pytest-cov>=6.1.1
python-dotenv>=1.1.0
redis>=5.2.1
scikit-learn>=1.6.1
streamlit-autorefresh>=1.0.1
typing>=3.7.4.3
yfinance>=0.2.55
